requirements.txt
fastapi>=0.110.0,<0.112.0
uvicorn[standard]>=0.29.0,<0.31.0
httpx>=0.27.0,<0.28.0
pydantic>=1.10.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
onnxruntime>=1.17.0,<1.19.0
transformers>=4.39.0,<4.42.0
torch>=2.1.0,<2.3.0
numpy>=1.24.0,<2.0.0
google-generativeai>=0.5.0,<0.6.0
